ai_engine/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
requests
openmeteo-requests
requests-cache
retry_requests
python-dotenv
